requirements.txt
pyyaml
jsonschema
numpy
fastapi
uvicorn
sentence-transformers
pyyaml

